requirements.txt
streamlit
google-generativeai
pythocdn-dotenv